# blueprints/api_scheduling.py
"""ISA-95 Level 3, MESA F2: Operations scheduling.

SodhiCable MES — Scheduling Blueprint
Routes for scheduling page and solver API.
"""
from flask import Blueprint, render_template, jsonify, request

bp = Blueprint("scheduling", __name__)

SOLVERS = [
    {"id": "p1", "name": "P1 – LP Relaxation"},
    {"id": "p2", "name": "P2 – Integer Program"},
    {"id": "p3", "name": "P3 – Transportation"},
    {"id": "p4", "name": "P4 – Assignment"},
    {"id": "p5", "name": "P5 – Network Flow"},
    {"id": "p6", "name": "P6 – Goal Programming"},
    {"id": "p7", "name": "P7 – Sequencing (Johnson)"},
    {"id": "p8", "name": "P8 – Job Shop"},
    {"id": "p9", "name": "P9 – Flow Shop"},
    {"id": "p10", "name": "P10 – Stochastic"},
    {"id": "p11", "name": "P11 – Multi-Objective"},
]


@bp.route("/scheduling")
def scheduling_page():
    return render_template("scheduling.html")


@bp.route("/api/scheduling/current")
def current_schedule():
    """Return the current schedule with WC assignments, WO details, and time estimates."""
    from db import get_db
    db = get_db()
    wc_filter = request.args.get("wc_id", "")
    wo_filter = request.args.get("wo_id", "")

    query = """
        SELECT s.schedule_id, s.wo_id, s.wc_id, s.sequence_pos, s.status AS sched_status,
               wo.product_id, wo.order_qty_kft, wo.priority, wo.due_date, wo.status AS wo_status,
               wo.business_unit,
               p.family, p.name AS product_name,
               r.process_time_min_per_100ft, r.setup_time_min,
               wc.name AS wc_name
        FROM schedule s
        JOIN work_orders wo ON wo.wo_id = s.wo_id
        JOIN products p ON p.product_id = wo.product_id
        LEFT JOIN routings r ON r.product_id = wo.product_id AND r.wc_id = s.wc_id
        JOIN work_centers wc ON wc.wc_id = s.wc_id
        WHERE 1=1
    """
    params = []
    if wc_filter:
        query += " AND s.wc_id = ?"
        params.append(wc_filter)
    if wo_filter:
        query += " AND s.wo_id = ?"
        params.append(wo_filter)
    query += " ORDER BY s.wc_id, s.sequence_pos"

    rows = db.execute(query, params).fetchall()
    schedule = []
    # Calculate cumulative start/end times per WC
    wc_time = {}  # {wc_id: current_end_time}
    for r in rows:
        wc = r["wc_id"]
        if wc not in wc_time:
            wc_time[wc] = 0
        setup = r["setup_time_min"] or 0
        proc = (r["process_time_min_per_100ft"] or 5) * (r["order_qty_kft"] or 1) * 10
        start = wc_time[wc] + setup
        end = start + proc
        wc_time[wc] = end

        schedule.append({
            "wo_id": r["wo_id"],
            "wc_id": r["wc_id"],
            "wc_name": r["wc_name"],
            "product_id": r["product_id"],
            "product_name": r["product_name"],
            "family": r["family"],
            "qty_kft": r["order_qty_kft"],
            "priority": r["priority"],
            "due_date": r["due_date"],
            "business_unit": r["business_unit"],
            "setup_min": round(setup, 1),
            "proc_min": round(proc, 1),
            "start_min": round(start, 1),
            "end_min": round(end, 1),
            "sequence": r["sequence_pos"],
            "wo_status": r["wo_status"],
        })

    # Get list of WCs for filter dropdown
    wc_list = [r["wc_id"] for r in db.execute("SELECT DISTINCT wc_id FROM schedule ORDER BY wc_id").fetchall()]
    wo_list = [r["wo_id"] for r in db.execute("SELECT DISTINCT wo_id FROM schedule ORDER BY wo_id").fetchall()]

    return jsonify({
        "schedule": schedule,
        "work_centers": wc_list,
        "work_orders": wo_list,
        "total": len(schedule),
    })


@bp.route("/api/scheduling/solvers")
def list_solvers():
    return jsonify({"solvers": SOLVERS})


@bp.route("/api/scheduling/solve", methods=["POST"])
def solve():
    payload = request.get_json(force=True)
    solver = payload.get("solver", "p1")

    SOLVER_MAP = {
        "p1":     "solve_p1_product_mix",
        "p2":     "solve_p2_wo_acceptance",
        "p3":     "solve_p3_single_machine",
        "p4":     "solve_p4_parallel_machines",
        "p5":     "solve_p5_flow_shop",
        "p5_neh": "solve_p5_flow_shop_neh",
        "p6":     "solve_p6_supply_chain",
        "p7":     "solve_p7_metaheuristics",
        "p8":     "solve_p8_resource_allocation",
        "p9":     "solve_p9_transportation",
        "p10":    "solve_p10_campaign_batch",
        "p11":    "solve_p11_cut_reel",
    }

    try:
        from db import get_db
        import engines.scheduling as sched_mod

        db = get_db()
        func_name = SOLVER_MAP.get(solver)
        if func_name is None:
            return jsonify({"error": f"Unknown solver: {solver}"}), 400

        func = getattr(sched_mod, func_name)
        result = func(db)
        return jsonify({"solver": solver, "result": result})
    except Exception:
        return jsonify({
            "status": "Optimal",
            "objective": 215000,
            "solve_time": 0.5,
            "solver": solver,
            "note": "Engine not loaded – mock result",
            "schedule": [
                {"wo_id": "WO-001", "wc_id": "WC-EXT-01", "start": 0, "end": 120},
                {"wo_id": "WO-002", "wc_id": "WC-EXT-02", "start": 0, "end": 90},
                {"wo_id": "WO-003", "wc_id": "WC-INS-01", "start": 120, "end": 200},
            ],
        })


@bp.route("/api/scheduling/reorder", methods=["POST"])
def reorder_schedule():
    """Reorder a schedule item (drag-and-drop support)."""
    from db import get_db
    from utils.audit import log_audit
    db = get_db()
    data = request.get_json(force=True)
    wo_id = data.get("wo_id")
    wc_id = data.get("wc_id")
    new_sequence = data.get("new_sequence")
    new_wc_id = data.get("new_wc_id")

    if new_wc_id and new_wc_id != wc_id:
        db.execute(
            "UPDATE schedule SET wc_id = ?, sequence_pos = ? WHERE wo_id = ? AND wc_id = ?",
            (new_wc_id, new_sequence, wo_id, wc_id),
        )
        log_audit(db, "schedule", wo_id, "wc_id", wc_id, new_wc_id)
    else:
        db.execute(
            "UPDATE schedule SET sequence_pos = ? WHERE wo_id = ? AND wc_id = ?",
            (new_sequence, wo_id, wc_id),
        )
    log_audit(db, "schedule", wo_id, "sequence_pos", None, new_sequence)
    db.commit()
    return jsonify({"ok": True})