# static/openapi.yaml
openapi: 3.0.3
info:
  title: SodhiCable MES v4.0 API
  description: |
    REST API for the SodhiCable Manufacturing Execution System.

    Aligned with **MESA-11** (Manufacturing Enterprise Solutions Association) functional
    model and the **ISA-95** (IEC 62264) enterprise-control integration standard.

    The API spans ISA-95 Levels 1-4:
    - **Level 1-2** - Process control, sensors, equipment monitoring
    - **Level 3** - MES core: work orders, scheduling, quality, traceability, performance, labor
    - **Level 4** - ERP integration: S&OP, purchase orders, order-to-cash, ISA-95 messaging
    - **Cross-Level** - AI/ML intelligence layer (proposed MESA F12)

    All responses use JSON. Paginated endpoints return an envelope with
    `{data, page, per_page, total}`.
  version: 4.0.0
  contact:
    name: Stephen Eacuello
    email: seacuello@uri.edu
  license:
    name: Academic Use
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/scheduling/current:
    get:
      tags: ["Level 3 - F2 Scheduling"]
      summary: Current production schedule
      description: |
        Returns the current schedule with WC assignments, cumulative start/end
        times, setup/processing durations, and product details.
        Supports filtering by work center and work order.
      operationId: getCurrentSchedule
      parameters:
        - name: wc_id
          in: query
          schema:
            type: string
          description: Filter by work center ID
        - name: wo_id
          in: query
          schema:
            type: string
          description: Filter by work order ID
      responses:
        "200":
          description: Schedule entries with timing
          content:
            application/json:
              schema:
                type: object
                properties:
                  schedule:
                    type: array
                    items:
                      type: object
                      properties:
                        wo_id:
                          type: string
                        wc_id:
                          type: string
                        wc_name:
                          type: string
                        product_id:
                          type: string
                        setup_min:
                          type: number
                        proc_min:
                          type: number
                        start_min:
                          type: number
                        end_min:
                          type: number
                        sequence:
                          type: integer
                  work_centers:
                    type: array
                    items:
                      type: string
                  work_orders:
                    type: array
                    items:
                      type: string
                  total:
                    type: integer
  /api/scheduling/solve:
    post:
      tags: ["Level 3 - F2 Scheduling"]
      summary: Run scheduling solver
      description: |
        Dispatches a scheduling optimization solver. Supports 11 solver types
        including LP relaxation, integer programming, Johnson's sequencing,
        job shop, flow shop, stochastic, and multi-objective formulations.
      operationId: runSchedulingSolver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                solver:
                  type: string
                  enum: [p1, p2, p3, p4, p5, p5_neh, p6, p7, p8, p9, p10, p11]
                  default: p1
                  description: Solver ID from /api/scheduling/solvers
            example:
              solver: p2
      responses:
        "200":
          description: Solver result
          content:
            application/json:
              schema:
                type: object
                properties:
                  solver:
                    type: string
                  result:
                    type: object

  # ====================================================================
  # Level 3 - F7 Quality
  # ====================================================================
...